GH-1515 moved inferencing tests out of compliance
