--- compliance/store/src/test/java/org/eclipse/rdf4j/sail/memory/MemInferencingTest.java
@@ -1,24 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2015 Eclipse RDF4J contributors, Aduna, and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Distribution License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/org/documents/edl-v10.php.
- *******************************************************************************/
-package org.eclipse.rdf4j.sail.memory;
-
-import org.eclipse.rdf4j.repository.Repository;
-import org.eclipse.rdf4j.repository.sail.SailRepository;
-import org.eclipse.rdf4j.sail.InferencingTest;
-import org.eclipse.rdf4j.sail.Sail;
-import org.eclipse.rdf4j.sail.inferencer.fc.ForwardChainingRDFSInferencer;
-import org.eclipse.rdf4j.sail.inferencer.fc.SchemaCachingRDFSInferencer;
-
-public class MemInferencingTest extends InferencingTest {
-
-	@Override
-	protected Repository createRepository() {
-		Sail sailStack = new SchemaCachingRDFSInferencer(new MemoryStore());
-		return new SailRepository(sailStack);
-	}
-}
