+++ core/sail/api/src/test/resources/testcases/rdf-mt-inferencing/domain/error001-in.nt
