+++ core/sail/api/src/test/resources/testcases/rdf-mt-inferencing/domain/test001-in.nt
