+++ core/sail/api/src/test/resources/testcases/rdf-mt-inferencing/range/test001-in.nt
