+++ core/sail/api/src/test/resources/testcases/rdf-mt-inferencing/subclassof/error001-out.nt
