+++ core/sail/api/src/test/resources/testcases/rdf-mt-inferencing/subclassof/test001-in.nt
