+++ core/sail/api/src/test/resources/testcases/rdf-mt-inferencing/subclassof/test002-out.nt
