+++ core/sail/api/src/test/resources/testcases/rdf-mt-inferencing/subclassof/test003-in.nt
