+++ core/sail/api/src/test/resources/testcases/rdf-mt-inferencing/subpropertyof/error001-in.nt
