+++ core/sail/api/src/test/resources/testcases/rdf-mt-inferencing/subpropertyof/test001-out.nt
