+++ core/sail/api/src/test/resources/testcases/rdf-mt-inferencing/subpropertyof/test002-out.nt
