+++ core/sail/api/src/test/resources/testcases/rdf-mt-inferencing/subpropertyof/test003-in.nt
