+++ core/sail/api/src/test/resources/testcases/rdf-mt-inferencing/subpropertyof/test004-out.nt
