+++ core/sail/api/src/test/resources/testcases/rdf-mt-inferencing/type/error001-in.nt
