+++ core/sail/api/src/test/resources/testcases/rdf-mt-inferencing/type/error002-in.nt
