+++ core/sail/api/src/test/resources/testcases/rdf-mt-inferencing/type/test001-out.nt
