+++ core/sail/api/src/test/resources/testcases/rdf-mt-inferencing/type/test002-out.nt
