+++ core/sail/api/src/test/resources/testcases/rdf-mt-inferencing/type/test003-out.nt
