+++ core/sail/api/src/test/resources/testcases/rdf-mt-inferencing/type/test004-out.nt
