+++ core/sail/api/src/test/resources/testcases/rdf-mt-inferencing/type/test005-out.nt
