+++ core/sail/api/src/test/resources/testcases/rdf-mt-inferencing/type/test006-out.nt
