+++ core/sail/inferencer/pom.xml
@@ -96,6 +96,12 @@
 			<version>${project.version}</version>
 			<scope>test</scope>
 		</dependency>
+				<dependency>
+			<groupId>${project.groupId}</groupId>
+			<artifactId>rdf4j-sail-nativerdf</artifactId>
+			<version>${project.version}</version>
+			<scope>test</scope>
+		</dependency>
 		<dependency>
 			<groupId>${project.groupId}</groupId>
 			<artifactId>rdf4j-sail-memory</artifactId>
