+++ core/sail/inferencer/src/test/java/org/eclipse/rdf4j/sail/inferencer/fc/MemInferencingTest.java
@@ -1,11 +1,11 @@
 /*******************************************************************************
- * Copyright (c) 2016 Eclipse RDF4J contributors.
+ * Copyright (c) 2015 Eclipse RDF4J contributors, Aduna, and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Distribution License v1.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/org/documents/edl-v10.php.
  *******************************************************************************/
-package org.eclipse.rdf4j.sail.memory;
+package org.eclipse.rdf4j.sail.inferencer.fc;
 
 import static junit.framework.TestCase.assertFalse;
 import static junit.framework.TestCase.assertTrue;
@@ -23,16 +23,16 @@ import org.eclipse.rdf4j.repository.Repository;
 import org.eclipse.rdf4j.repository.RepositoryConnection;
 import org.eclipse.rdf4j.repository.sail.SailRepository;
 import org.eclipse.rdf4j.sail.InferencingTest;
-import org.eclipse.rdf4j.sail.inferencer.fc.SchemaCachingRDFSInferencer;
+import org.eclipse.rdf4j.sail.Sail;
+import org.eclipse.rdf4j.sail.memory.MemoryStore;
 import org.junit.Test;
 
-public class SchemaCachingRDFSInferencerMemInferencingTest extends InferencingTest {
+public class MemInferencingTest extends InferencingTest {
 
 	@Override
-	protected Repository createRepository() {
-		SchemaCachingRDFSInferencer sailStack = new SchemaCachingRDFSInferencer(new MemoryStore(), true);
-		// sailStack.setAddInferredStatementsToDefaultContext(false);
-		return new SailRepository(sailStack);
+	protected Sail createSail() {
+		Sail sailStack = new SchemaCachingRDFSInferencer(new MemoryStore());
+		return sailStack;
 	}
 
 	@Test
@@ -69,8 +69,8 @@ public class SchemaCachingRDFSInferencerMemInferencingTest extends InferencingTe
 
 	@Test
 	public void testBlankNodePredicateInference() {
-		Repository sailRepository = createRepository();
-		sailRepository.initialize();
+		Repository sailRepository = new SailRepository(createSail());
+
 		ValueFactory vf = sailRepository.getValueFactory();
 
 		try (RepositoryConnection connection = sailRepository.getConnection()) {
@@ -92,8 +92,7 @@ public class SchemaCachingRDFSInferencerMemInferencingTest extends InferencingTe
 
 	@Test
 	public void testRollback() throws NoSuchMethodException, InvocationTargetException, IllegalAccessException {
-		Repository sailRepository = createRepository();
-		sailRepository.initialize();
+		Repository sailRepository = new SailRepository(createSail());
 		ValueFactory vf = sailRepository.getValueFactory();
 
 		IRI A = vf.createIRI("http://A");
@@ -140,8 +139,7 @@ public class SchemaCachingRDFSInferencerMemInferencingTest extends InferencingTe
 
 	@Test
 	public void testFastInstantiate() throws NoSuchMethodException, InvocationTargetException, IllegalAccessException {
-		Repository sailRepository = createRepository();
-		sailRepository.initialize();
+		Repository sailRepository = new SailRepository(createSail());
 		ValueFactory vf = sailRepository.getValueFactory();
 
 		IRI A = vf.createIRI("http://A");
@@ -169,5 +167,4 @@ public class SchemaCachingRDFSInferencerMemInferencingTest extends InferencingTe
 					correctInference);
 		}
 	}
-
 }
